Language: optimisations and logic fixes to batching
- Dealing with special cases for the batch statement. For example, if the batch statement ends its first path without queueing anything (20/12/2021 - 22:30:38)
- Also dealt with anytimes there might be a mismatch of method calls within the batch (20/12/2021 - 22:39:26)
- Modified example 11 to showcase these new changes (20/12/2021 - 22:45:21)
- Changed BatchResult implementation to include a parser.MethodCall pointer as well a getter for it (20/12/2021 - 22:48:04)
- Had to convert some values in returned constructed value from Method.Call to be float64 rather than int (20/12/2021 - 22:49:57)

+++ src/eval/methods.go
@@ -187,10 +187,10 @@ func (m *Method) Call(args ...*data.Value) (err error, value *data.Value) {
 						cookies[i] = map[string]interface{}{
 							"name": cookie.Name,
 							"value": cookie.Value,
-							"max_age": cookie.MaxAge,
+							"max_age": float64(cookie.MaxAge),
 							"secure": cookie.Secure,
 							"http_only": cookie.HttpOnly,
-							"same_site": cookie.SameSite,
+							"same_site": float64(cookie.SameSite),
 							"raw": cookie.Raw,
 						}
 					}
@@ -204,9 +204,9 @@ func (m *Method) Call(args ...*data.Value) (err error, value *data.Value) {
 					return headers
 				}(),
 				"received": resp.ReceivedAt().String(),
-				"size": resp.Size(),
+				"size": float64(resp.Size()),
 				"status": resp.Status(),
-				"code": resp.StatusCode(),
+				"code": float64(resp.StatusCode()),
 				"time": resp.Time().String(),
 			},
 			Type:     data.Object,
